requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pillow==10.0.1

# Optional dependencies for emotion detection
tensorflow==2.13.0
# or tensorflow-cpu==2.13.0 for CPU-only version

# Additional utilities
python-dateutil==2.8.2
requests==2.31.0